Sell crypto: ref models and moonpay models
